+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [main]
+## [3.11.2] - 2025-05-21
+
+### Fixed
+
+- Images and videos that are being used as source material for light effects
+  are now properly converted from sRGB to linear color space, thanks to
+  @flopbuster
+
+## [3.11.1] - 2025-05-06
 
 ### Fixed